membuat fitur komplen

+++ ProjekUasFpwKel5_SistemLaundry/database/migrations/2025_11_10_201648_create_riwayats_table.php
@@ -11,15 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('riwayats', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('pesanan_id');
-            $table->enum('tipe', ['Pemasukan', 'Pengeluaran']);
-            $table->decimal('jumlah', 10, 2);
-            $table->timestamps();
+        if (!Schema::hasTable('riwayats')) {
+            Schema::create('riwayats', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('pesanan_id');
+                $table->enum('tipe', ['Pemasukan', 'Pengeluaran']);
+                $table->decimal('jumlah', 10, 2);
+                $table->timestamps();
 
-            $table->foreign('pesanan_id')->references('id')->on('pesanans')->onDelete('cascade');
-        });
+                $table->foreign('pesanan_id')->references('id')->on('pesanans')->onDelete('cascade');
+            });
+        }
     }
 
     /**
